backend/requirements.txt
affine==2.4.0
aioredis==2.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asciitree==0.3.3
async-timeout==5.0.1
attrs==25.4.0
bioio==3.0.0
bioio-base==3.0.0
bioio-ome-tiff==1.4.0
certifi==2025.11.12
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.2
colorcet==3.1.0
contourpy==1.3.2
cycler==0.12.1
dask==2025.11.0
distributed==2025.11.0
einops==0.8.1
elementpath==5.0.4
exceptiongroup==1.3.0
fastapi==0.121.2
fasteners==0.20
fastremap==1.17.7
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
geojson==3.2.0
h11==0.16.0
idna==3.11
imagecodecs==2025.3.30
ImageIO==2.37.2
importlib_metadata==8.7.0
instanseg-torch @ git+https://github.com/instanseg/instanseg.git@3470dba7cf56a7510de172857f7d7063cdb4fa2d
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
locket==1.0.0
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numcodecs==0.13.1
numpy==1.26.4
ome-types==0.6.2
opencv-python==4.11.0.86
openslide-python==1.4.2
packaging==25.0
palettable==3.3.3
pandas==2.3.3
partd==1.4.2
pillow==12.0.0
psutil==7.1.3
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rasterio==1.4.3
redis==7.0.1
scikit-image==0.25.2
scipy==1.15.3
semver==3.0.4
six==1.17.0
slideio==2.7.3
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
sympy==1.14.0
tblib==3.2.2
tifffile==2024.12.12
tiffslide==2.5.1
toolz==1.1.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xarray==2025.6.1
xmlschema==4.2.0
xsdata==25.7
zarr==2.18.3
zict==3.0.0
zipp==3.23.0
